refactor: CashLogService에서 도메인 서비스 로직 분리
