+++ back/src/main/java/com/ll/goohaeyou/jobPost/employ/domain/service/EmployDomainService.java
@@ -4,8 +4,10 @@ import com.ll.goohaeyou.global.standard.anotations.DomainService;
 import com.ll.goohaeyou.jobApplication.domain.type.WageStatus;
 import com.ll.goohaeyou.jobPost.employ.exception.EmployException;
 import com.ll.goohaeyou.jobPost.jobPost.domain.type.WagePaymentMethod;
+import org.springframework.transaction.annotation.Transactional;
 
 @DomainService
+@Transactional(readOnly = true)
 public class EmployDomainService {
 
     public WageStatus determineWageStatus(WagePaymentMethod wagePaymentMethod) {
