+++ back/src/main/java/com/ll/goohaeyou/payment/payment/application/PaymentInfoService.java
@@ -1,7 +1,7 @@
 package com.ll.goohaeyou.payment.payment.application;
 
 import com.ll.goohaeyou.payment.payment.application.dto.PaymentInfoResponse;
-import com.ll.goohaeyou.payment.payment.domain.PaymentDomainService;
+import com.ll.goohaeyou.payment.payment.domain.service.PaymentDomainService;
 import com.ll.goohaeyou.payment.payment.domain.entity.Payment;
 import com.ll.goohaeyou.payment.payment.domain.policy.PaymentPolicy;
 import lombok.RequiredArgsConstructor;
